third_party/blink/web_tests/external/wpt/wasm/jsapi/memory/to-resizable-buffer.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] API surface
  assert_equals: expected "object" but got "undefined"
[FAIL] toResizableBuffer caching behavior
  memory.toResizableBuffer is not a function
[FAIL] toResizableBuffer max size
  memory.toResizableBuffer is not a function
[FAIL] Resizing a Memory's resizable buffer
  memory.toResizableBuffer is not a function
[FAIL] Resizable buffers from Memory cannot be detached by JS
  memory.toResizableBuffer is not a function
Harness: the test ran to completion.
